Исправлены замечания и добавлены комментарии
